Change type from size_t to int to save space

+++ llmc/rand.h
@@ -220,14 +220,14 @@ void normal_(float* data, unsigned int numel, float mean, float std, mt19937_sta
     }
 }
 
-void random_permutation_with_init(size_t* data, unsigned int numel, mt19937_state* state) {
-    for (unsigned int i = 0; i < numel; i++) {
+void random_permutation_with_init(int* data, unsigned int numel, mt19937_state* state) {
+    for (int i = 0; i < numel; i++) {
         data[i] = i;
     }
-    for (unsigned int i = numel - 1; i > 0; i--) {
-        unsigned int j = randint32(state) % (i + 1);
+    for (int i = numel - 1; i > 0; i--) {
+        int j = randint32(state) % (i + 1);
         // swap
-        unsigned int tmp = data[i];
+        int tmp = data[i];
         data[i] = data[j];
         data[j] = tmp;
     }
